Crate tokyodoves

Crate Description

Re-exports

pub use array_macro;

pub use strum;

pub use strum_macros;

pub use thiserror;

Modules

analysis

Convenient tools to analyze the game

error

Error variants

game

Structs

ActionsBwd

An ActionContainer returned by Board::legal_actions_bwd

ActionsBwdIntoIter

An Iterator returned by ActionsBwd::into_iter

ActionsBwdIter

An Iterator returned by ActionsBwd::iter

ActionsFwd

An ActionContainer returned by Board::legal_actions

ActionsFwdIntoIter

An Iterator returned by ActionsFwd::into_iter

ActionsFwdIter

An Iterator returned by ActionsFwd::iter

Board

An implementation of Tokyo Doves board based on bitboard techniques

BoardBuilder

A builder of Board

BoardDisplay

ColorIter

An iterator over the variants of Color

DoveIter

An iterator over the variants of Dove

DoveSet

Read-only Sized container of Doves without duplication.

DoveSetIntoIter

An Iterator returned by DoveSet::into_iter

Rectangle

Rectangle with edges of usize coordinates.

Shift

Difference between two squares.

Enums

Action

Actions players can perform in their turn.

BoardDisplayFormat

Color

Two colors of player, just like black and white in chess.

Dove

Six types of doves.

SurroundedStatus

An enum returned by Board::surrounded_status

Traits

ActionContainer

Read-only Sized container of Actions.

Functions

color_dove_to_char

try_char_to_color_dove