Module tokyodoves::analysis

Convenient tools to analyze the game.

The following functions are main contents:

evaluate_board
It calculates a possible range of BoardValue of specified Board.

compare_board_value
It compares the value of specified Board to a given BoardValue.

create_checkmate_tree and create_checkmate_tree_with_value
It creates an BoardValueTree that describes routes to ends of the game.

find_best_actions
It collects the best Actions by BoardValue.

The BoardValue struct plays important roles in the all above functions. See their documentations for more.

Structs

BoardValue

Value of Board.

BoardValueTree

A tree structure that contains Board and BoardValue on its nodes, and Actions on its edges.

Interval

A struct representing closed interval between two BoardValues.

PositionMapper

A mapping of positions used to canonicalize functions of a Board.

TreeDisplay

A struct to configure display styles of BoardValueTree.

Enums

BoardValueKind

Kinds of BoardValue

TreeDisplayFormat

Formats of display used by TreeDisplay.

Functions

compare_board_value

Compares the value of specified Board to a given BoardValue.

create_checkmate_tree

Creates an BoardValueTree that describes routes to ends of the game.

create_checkmate_tree_with_value

Creates an BoardValueTree that describes routes to ends of the game.

evaluate_board

Calculates a possible range of BoardValue of specified Board.

find_best_actions

Collects the best Actions by BoardValue.