Function tokyodoves::analysis::compare_board_value

pub fn compare_board_value(
    board: Board,
    value: BoardValue,
    player: Color,
    rule: GameRule
) -> Result<Ordering, Error>

Compares the value of specified Board to a given BoardValue.

It returns:

Ok(Greater) if the value of board is greater than value

Ok(Equal) if the value of board equals to value

Ok(Less) if the value of board is less than `value

Errors

Returns Err only when the argument is invalid. Specifically, the following cases are invalid:

board is already finished (at least one boss is surrounded)

value is Unknown, Win(even number) or Lose(odd or zero)

rule.suicide_atk_judge is Judge::Draw