Function tokyodoves::analysis::create_checkmate_tree_with_value

pub fn create_checkmate_tree_with_value(
    board: Board,
    value: BoardValue,
    player: Color,
    rule: GameRule
) -> Result<BoardValueTree, Error>

Creates an BoardValueTree that describes routes to ends of the game.

This function behaves almost similar to create_checkmate_tree, except that it receives value (BoardValue) instead of search depth, and requires value coinsides with the value of board. This function runs much faster than create_checkmate_tree_with_value in exchange for restriction of possible arguments.

Errors

Returns Err when the argument is invalid. Specifically, the following cases are invalid:

board is already finished (at least one boss is surrounded)

value is Unknown, Win(even number) or Lose(odd or zero)

rule.suicide_atk_judge is Judge::Draw Furthermore, it returns Err when the value of board differs from value. In such a case, returned value contains a variant of std::cmp::Ordering, which means that the value of board is greater (Greater) or less (Less) than the specified value.