Function tokyodoves::analysis::evaluate_board

pub fn evaluate_board(
    board: Board,
    player: Color,
    search_depth: usize,
    rule: GameRule
) -> Result<Interval, Error>

Calculates a possible range of BoardValue of specified Board.

It searches the value of board by pursuing search_depth turns forward. The result is returned in Interval, which is a closed interval between two BoardValues. It means that the value of the board is in the interval.

Errors

Returns Err only when the argument is invalid. Specifically, the following cases are invalid:

board is already finished (at least one boss is surrounded)

rule.suicide_atk_judge is Judge::Draw