Struct tokyodoves::analysis::BoardValueTree

pub struct BoardValueTree { /* private fields */ }

A tree structure that contains Board and BoardValue on its nodes, and Actions on its edges.

It is returned by create_checkmate_tree.

Implementations

impl BoardValueTree

pub fn board(&self) -> Board

Returns Board at the root node.

pub fn player(&self) -> &Color

Returns a reference to the player Color at the root node.

pub fn value(&self) -> &BoardValue

Returns a reference to the BoardValue at the root node.

pub fn display(&self) -> TreeDisplay<'_>

Returns TreeDisplay to display self.

pub fn child(&self, action: &Action) -> Option<&BoardValueTree>

Returns a reference to a child tree associated to the specified Action.

It returns Some(&child) if a child exists, otherwise None.

pub fn actions(&self) -> Keys<'_, Action, BoardValueTree>

Returns an Iterator that iterates over all Actions connecting to children.

pub fn children(&self) -> Values<'_, Action, BoardValueTree>

Returns an Iterator that iterates over all children.

pub fn num_children(&self) -> usize

Counts the number of children.

pub fn is_leaf(&self) -> bool

Returns true if it is a leaf node, i.e., it has no children.

pub fn actions_children(&self) -> Iter<'_, Action, BoardValueTree>

Returns an Iterator that iterates all pairs of Action and BoardValueTree.

pub fn depth(&self) -> usize

Returns the depth of the tree, i.e., the longest steps from the root node to leaf node.

pub fn is_good_for_puzzle(&self, step: usize) -> bool

Returns true if the tree is a single road during the first step steps.

pub fn save_as_dot<W>(&self, writer: W) -> Result<()>where W: Write,

Saves the tree as a file in DOT language.

DOT language is a graph description language, used in a graph visualization software Graphviz.

Examples

The saved file can be converted to colorful graph by the following command:

graphviz.exe -Tsvg tree.dot -o output.svg

See the documentation of Graphviz for more.