Module tokyodoves::collections

A module containing a light Board container BoardSet and associated items.

Re-exports

pub use board_set::BoardSet;

pub use board_set::Capacity;

Modules

board_set

A module containing a light Board container BoardSet and associated items.

Structs

LazyBoardLoader

A utility to load Boards in a lazy way from the binary file saved by the save method of BoardSet.

LazyRawBoardLoader

A struct almost the same as LazyBoardLoader, except that it loads u64 expressions of Boards instead.