Struct tokyodoves::collections::board_set::Capacity

pub struct Capacity(/* private fields */);

A capacity, i.e., what size of memory is allocated by BoardSet or RawBoardSet.

Note that, unlike HashSet, the capacity does not behaves like usize. It has an internal data HashMap<u32, usize>, where keys represents top half of u64 expression of Boards and values represents how many elements the set can hold. An addition of two capacities is defined by the addition of values of internal hash maps sharing the same keys.

Examples

use tokyodoves::Board;
use tokyodoves::collections::BoardSet;

let mut set = BoardSet::new();
set.insert(Board::new());
let capacity = set.capacity();

Implementations

impl Capacity

pub fn new() -> Self

Returns an empty capacity.

Examples

use tokyodoves::collections::{BoardSet, Capacity};

let empty = BoardSet::new().capacity();
assert_eq!(empty, Capacity::new());

pub fn len(&self) -> usize

Returns the number of Boards (or u64s) the BoardSet (or RawBoardSet) with the capacity can hold.

Examples

use tokyodoves::Board;
use tokyodoves::collections::BoardSet;

let mut set = BoardSet::new();
set.insert(Board::new());
assert!(set.capacity().len() >= 1);

pub fn is_empty(&self) -> bool

Returns true if no capacity.

Examples

use tokyodoves::Board;
use tokyodoves::collections::BoardSet;

let mut set = BoardSet::new();
assert!(set.capacity().is_empty());
set.insert(Board::new());
assert!(!set.capacity().is_empty());