Module tokyodoves::error

Error variants

A rough sketch of dependency tree:

ActionConvertError

SSNEncodingErrorType

SSNDecodingErrorType

BoardError

ActionPerformErrorType

BoardCreateErrorType

[GameError]

BoardError

…

GameRuleError

Enums

ActionConvertError

Errors on conversion between Action and string in SSN

ActionPerformErrorType

Types of errors on performing Action

BoardCreateErrorType

Types of errors on creating Board

BoardError

Errors associated to Board

GameRuleError

Errors associated to GameRule

SSNDecodingErrorType

Types of errors on conversion of Action from string in SSN

SSNEncodingErrorType

Types of errors on conversion of Action into string in SSN