Function tokyodoves::try_char_to_color_dove

pub fn try_char_to_color_dove(c: char) -> Option<(Color, Dove)>

Converts char to a pair of Color and Dove.

The mapping rule is shown in the table below:

Dove \ Color	Red	Green

B	‘B’	‘b’

A	‘A’	‘a’

Y	‘Y’	‘y’

M	‘M’	‘m’

T	‘T’	‘t’

H	‘H’	‘h’

It returns Some((Color, Dove)) when an appropreate pair is found, otherwise returns None.