Module tokyodoves::game

Convenient tools for playing games

Main contents:

Game
A struct that provides methods to play the game.

Agent
A trait to define an agent playing the game. The following agents are included:

RandomAgent
An agent who chooses the next action at random.

AnalystAgent
An agent who chooses the next action based on some analyses.

ConsoleAgent
An agent who asks the next action to the console. It provides CLI with humans.

You can define another agent in you code by implementing the Agent trait if you want.

Arena
A struct of an arena, where two Agents play against.

Structs

AnalystAgent

An Agent who choses a next action based on analysis of the status.

Arena

A struct to let two Agents to play the game.

ConsoleAgent

An Agent who asks what action should be performed to the console.

Game

A struct that provides methods to play games.

GameDisplay

A struct to configure display styles of Game.

GameRule

A struct holding detailed rules for playing games.

RandomAgent

An Agent who chooses a next action at random.

Enums

GameDisplayFormat

Formats of display used by GameDisplay

GameStatus

Status of the game

Judge

Judgement of winner on some event

Traits

Agent

An agant who plays the game.