Struct tokyodoves::game::Game

pub struct Game { /* private fields */ }

A struct that provides methods to play Tokyo Doves games following rules.

Examples

The following is a simple example in which one game is played:

use tokyodoves::ActionContainer;
use tokyodoves::game::Game;

fn main() -> Result<(), Box<dyn std::error::Error>> {
    // Create a [`Game`] object allowing `Remove` action
    let mut game = Game::new(true);
    // Displays the status of the game
    println!("{}", game);

    let mut turn = 0;
    loop {
        turn += 1;
        // Collects all possible actions
        let actions = game.legal_actions();
        // Selects one of them by methods of ActionContainer trait
        let action = actions[turn % actions.len()];
        println!("  --> {}", action.to_ssn(game.board())?);

        game.perform(action)?; // Performs the selected action
        println!("{}", game);

        if !game.is_ongoing() { // If the game is not ongoing...
            break; // breaks the loop!
        }
    }

    match game.winner() { // Who won the game?
        Some(player) => println!("*** {:?} win! ***", player), // A player won the game
        None => println!("*** Draw! ***"), // It was a draw game
    }

    Ok(())
}

To customize the rule more, you can create Game from GameRule object:

use tokyodoves::Color;
use tokyodoves::game::{Game, GameRule};

let mut rule = GameRule::default().with_is_remove_accepted(false).with_first_player(Color::Green);
let game = Game::new_with_rule(rule);
println!("{}", game);

For more information about the default value of GameRule, see the documentation about the implementation of Default trait for GameRule

Implementations

impl Game

pub fn new(is_remove_accepted: bool) -> Game

Constructs Game

pub fn new_with_rule(rule: GameRule) -> Game

Constructs Game with a specified rule

pub fn reset(&mut self)

Reset Game to the initial state

pub fn rule(&self) -> &GameRule

Get a reference to game rule

pub fn board(&self) -> &Board

Get a reference to board

pub fn next_player(&self) -> &Color

Get a reference to the next player

pub fn status(&self) -> &GameStatus

Get a reference to status

pub fn is_ongoing(&self) -> bool

Returns true if the game is ongoing

pub fn winner(&self) -> Option<Color>

Returns winner.

It returns Some(Color) if winner has been determined. It returns None if the game is ongoing or draw.

pub fn legal_actions(&self) -> ActionsFwd

Returns an ActionContainer of legal Actions.

pub fn perform(&mut self, action: Action) -> Result<(), GameError>

Performs action.

Errors

It returns:

Err(GameError::GameFinishedError) if the game has already been finished.

Err(GameError::PlayerMismatchError) if the player of action is different from the next player.

Err(GameError::BoardError { .. }) if performing action is illegal for board.

In any cases, Game object is left unchanged.