Struct tokyodoves::Board

pub struct Board { /* private fields */ }

An implementation of Tokyo Doves board based on bitboard techniques

Implementations

impl Board

pub fn perform(&mut self, action: Action) -> Result<(), BoardError>

Performs the specified action to self.

Errors

Examples

use tokyodoves::{BoardBuilder, Action, Color, Dove, Shift};

let mut board = BoardBuilder::new().build()?;
let action = Action::Put(Color::Red, Dove::A, Shift { dh: 1, dv: 0 });
let result = board.perform(action);
assert!(result.is_ok());

pub fn perform_copied(&self, action: Action) -> Result<Self, BoardError>

Returns the result of performing the specified action to self. It differs from perform in that it does not change self.

Errors

Examples

use tokyodoves::{BoardBuilder, Action, Color, Dove, Shift};

let board = BoardBuilder::new().build()?;
let action = Action::Put(Color::Red, Dove::A, Shift { dh: 1, dv: 0 });
let next_board = board.perform_copied(action)?;

pub fn perform_bwd(&mut self, action: Action) -> Result<(), BoardError>

Performs the specified backward action to self.

Errors

Examples

use tokyodoves::{BoardBuilder, Action, Color, Dove, Shift};

let mut board = BoardBuilder::new().build()?;
let action = Action::Put(Color::Red, Dove::A, Shift { dh: 1, dv: -1 });
let result = board.perform_bwd(action);
assert!(result.is_ok());

pub fn perform_bwd_copied(&self, action: Action) -> Result<Self, BoardError>

Returns the result of performing the specified backward action to self. It differs from perform_bwd in that it does not change self.

Errors

Examples

use tokyodoves::{BoardBuilder, Action, Color, Dove, Shift};

let board = BoardBuilder::new().build()?;
let action = Action::Put(Color::Red, Dove::A, Shift { dh: 1, dv: 0 });
let next_board = board.perform_bwd_copied(action)?;

pub fn perform_unchecked(&mut self, action: Action) -> Result<(), BoardError>

Performs action to self (without distinction between forward and backward).

It may return Ok(()) even if the specified action is illegal. In such a case, the resulting board may violate the rule (e.g. some dove would be isolated from others). If you are not sure that the action is legal, consider to select one of the following choices:

Call check_action or check_action_bwd to check legality in advance.

Call perform or perform_bwd insted of this method, which calls check_action or check_action_bwd internally. This method is faster than perform and perform_bwd thanks to absence of legality check.

Errors

Returns Err(error::BoardError::MaskShiftError { .. }) if the target position of Action::Move or Action::Put is more than two squares far from the current 4x4 field (, which are clearly illegal). When it returns an error, self is not changed.

Examples

use tokyodoves::{BoardBuilder, Action, Color, Dove, Shift};

let mut board = BoardBuilder::new().build()?;
let action = Action::Put(Color::Red, Dove::A, Shift { dh: 1, dv: 0 });
let result = board.perform_unchecked(action);
assert!(result.is_ok());

pub fn perform_unchecked_copied( &self, action: Action ) -> Result<Self, BoardError>

Returns the result of performing the specified action to self. It differs from perform_unchecked in that it does not change self.

Errors

Returns Err(error::BoardError::MaskShiftError { .. }) if the target position of Action::Move or Action::Put is more than two squares far from the current 4x4 field (, which are clearly illegal).

Examples

use tokyodoves::{BoardBuilder, Action, Color, Dove, Shift};

let board = BoardBuilder::new().build()?;
let action = Action::Put(Color::Red, Dove::A, Shift { dh: 1, dv: 0 });
let next_board = board.perform_unchecked_copied(action)?;

pub fn check_action(&self, action: Action) -> Result<(), BoardError>

Check if action is legal.

Errors

Examples

use tokyodoves::{BoardBuilder, Action, Color, Dove, Shift};

let mut board = BoardBuilder::new().build()?;
let action = Action::Put(Color::Red, Dove::A, Shift { dh: 1, dv: 0 });
assert!(board.check_action(action).is_ok());

pub fn check_action_bwd(&self, action: Action) -> Result<(), BoardError>

Check if backward action is legal.

Errors

Examples

use tokyodoves::{BoardBuilder, Action, Color, Dove, Shift};

let mut board = BoardBuilder::new().build()?;
let action = Action::Put(Color::Red, Dove::A, Shift { dh: 1, dv: -1 });
assert!(board.check_action_bwd(action).is_ok());

pub fn legal_actions( &self, player: Color, contains_put: bool, contains_move: bool, contains_remove: bool ) -> ActionsFwd

Collects and returns the set of all legal Actions for the specified player. Three boolean arguments — contains_put, contains_move and contains_remove — controls what kinds of Actions should be contained.

Examples

use tokyodoves::{BoardBuilder, Color};

let board = BoardBuilder::new().build()?;
let actions = board.legal_actions(Color::Red, true, true, true);
for action in actions {
    println!("{}", board.perform_copied(action)?);
}

pub fn legal_actions_bwd( &self, player: Color, contains_put: bool, contains_move: bool, contains_remove: bool ) -> ActionsBwd

Collects and returns the set of all legal backward-Actions for the specified player.

Examples

use tokyodoves::{BoardBuilder, Color};

let board = BoardBuilder::new().build()?;
let actions = board.legal_actions_bwd(Color::Red, true, true, true);
for action in actions {
    println!("{}", board.perform_bwd_copied(action)?);
}

pub fn swap_color(&mut self)

Swaps the colors red and green.

pub fn position_in_rbcc(&self, player: Color, dove: Dove) -> Option<Shift>

Returns the position of specified player and dove in Red-Boss-Centered Coordinate (RBCC).

dv and dh in RBCC is shown in below (B is red boss):

    dh  -3   -2   -1    0    +1   +2   +3
  dv  +----+----+----+-----+----+----+----+
  -3  |    |    |    |     |    |    |    |
      +----+----+----+-----+----+----+----+
  -2  |    |    |    |     |    |    |    |
      +----+----+----+-----+----+----+----+
  -1  |    |    |    |     |    |    |    |
      +----+----+----+-----+----+----+----+
   0  |    |    |    |  B  |    |    |    |
      +----+----+----+-----+----+----+----+
  +1  |    |    |    |     |    |    |    |
      +----+----+----+-----+----+----+----+
  +2  |    |    |    |     |    |    |    |
      +----+----+----+-----+----+----+----+
  +3  |    |    |    |     |    |    |    |
      +----+----+----+-----+----+----+----+

pub fn count_doves_on_field(&self) -> usize

Counts the number of all doves on the field.

pub fn count_doves_on_field_of(&self, player: Color) -> usize

Counts the number of doves of the player on the field.

pub fn doves_in_hand_of(&self, player: Color) -> DoveSet

Collects and returns the set of doves in the player’s hand.

pub fn doves_on_field_of(&self, player: Color) -> DoveSet

Collects and returns the set of doves of the player on the field.

pub fn is_on_field(&self, player: Color, dove: Dove) -> bool

Returns true if the player’s dove is on the field.

pub fn is_in_hand(&self, player: Color, dove: Dove) -> bool

Returns true if the player’s dove is in their hand.

pub fn liberty(&self, player: Color, dove: Dove) -> Option<usize>

Returns “liberty” of player’s dove, where liberty is the number of free squares next to the specified dove.

Note that None and Some(0) have different meanings; None means that the specified dove is not on the board while Some(0) means that the specified dove is on the field and completely surrounded.

Examples

+---+---+---+---+
| b | T |   |   |
+---+---+---+---+
|   | B |   |   |
+---+---+---+---+
|   | a | H | M |
+---+---+---+---+
|   |   | m |   |
+---+---+---+---+

In this case, the liberty of “B” is two. The liberty of “M” is three, because two of four squares next to “M” is occupied by “H” and a wall.

pub fn liberty_of_boss(&self, player: Color) -> usize

Returns “liberty” of player’s boss-hato.

The meaning of liberty is written in liberty section. A player loses when the liberty of their boss-hato becomes 0.

Panics

Panics if player’s boss is not on the field, which does not happen in normal play.

pub fn surrounded_status(&self) -> SurroundedStatus

Get info whether bosses are surrounded or not

pub fn minimum_rectangle(&self) -> Rectangle

Returns the minimum rectangle that contains all doves on the field.

It has fields hmin, hmax, vmin and vmax, standing for coordinates of edges of the rectangle.

Examples

[Case 1]
    0   1   2   3
  +---+---+---+---+
0 | b |   | T |   |
  +---+---+---+---+
1 |   | B | a |   |
  +---+---+---+---+
2 |   |   |   |   |
  +---+---+---+---+
3 |   |   |   |   |
  +---+---+---+---+
-> hmin=0, hmax=2, vmin=0, vmax=1

[Case 2]
    0   1   2   3
  +---+---+---+---+
0 |   |   |   |   |
  +---+---+---+---+
1 |   | b |   |   |
  +---+---+---+---+
2 | a |   | H | B |
  +---+---+---+---+
3 |   |   |   |   |
  +---+---+---+---+
-> hmin=0, hmax=3, vmin=1, vmax=2

[Case 3]
    0   1   2   3
  +---+---+---+---+
0 |   |   |   |   |
  +---+---+---+---+
1 |   |   |   |   |
  +---+---+---+---+
2 |   |   |   |   |
  +---+---+---+---+
3 | h | B | b | Y |
  +---+---+---+---+
-> hmin=0, hmax=3, vmin=3, vmax=3

pub fn to_u64(&self) -> u64

Returns a light expression of 64 bits.

64 bits consist of the following three parts:

Top 4 bits: empty (filled by 0)

Next 12 bits: each bit means that each kind of dove is on the field (1) or not (0)

Next 48 bits: positions of doves are expressed by 0~15, which consumes 4 bits per one dove (4 bits x 12 doves = 48 bits)

The order of 12 doves is “B > b > A > a > Y > y > M > m > T > t > H > h”, where capital/lower cases mean red/green dove, respectively.

pub fn to_canonical_u64(&self, next_player: Color) -> u64

Returns a light expression of u64 with a universality.

The composition of the returned value is the same as that of to_u64. “Universality” means that this method returns the same value for boards that

coinside with each other under reflection, rotation, translation, and any compositions of them.

coinside with each other after alternating next player and swapping colors of doves simultaneously.

pub fn to_4x4_matrix(&self) -> [[Option<(Color, Dove)>; 4]; 4]

Returns 4x4 matrix (array of array) representing the board.

Examples

use tokyodoves::{BoardBuilder, Color, Dove};

let board = BoardBuilder::new().build()?;
let matrix = [
    [Some((Color::Green, Dove::B)), None, None, None],
    [Some((Color::Red, Dove::B)), None, None, None],
    [None, None, None, None],
    [None, None, None, None],
];
assert_eq!(board.to_4x4_matrix(), matrix);

pub fn to_framed_string(&self) -> String

Returns a String expression with a frame.

Examples

use tokyodoves::BoardBuilder;

let board = BoardBuilder::new().build()?;
let string: String = [
    "+---+---+---+---+",
    "| b |   |   |   |",
    "+---+---+---+---+",
    "| B |   |   |   |",
    "+---+---+---+---+",
    "|   |   |   |   |",
    "+---+---+---+---+",
    "|   |   |   |   |",
    "+---+---+---+---+",
].join("\n");
assert_eq!(board.to_framed_string(), string);