Struct tokyodoves::Shift

pub struct Shift { /* private fields */ }

A struct representing difference between two squares.

It has two data dh, i.e. the horizonal difference, and dv, i.e. the vertical difference. Positive/negative dh means the direction is on right/left. Positive/negative dv means the direction is on downwards/upwards. These data can be accessed via dh, dv, dh_mut and dv_mut.

This struct supports addition, subtraction and negation.

use tokyodoves::Shift;

assert_eq!(-Shift::new(2, 3), Shift::new(-2, -3));
assert_eq!(Shift::new(1, 1) + Shift::new(2, 0), Shift::new(3, 1));
assert_eq!(Shift::new(2, -1) - Shift::new(1, -2), Shift::new(1, 1));

Pedantically speaking, this struct behaves like an element in a two-dimensional Affine space with integral coordinates without scalar multiplication.

Implementations

impl Shift

pub fn new(dv: i8, dh: i8) -> Self

Constructs Shift.

pub fn dh(&self) -> &i8

Returns a reference to the vertical difference of coordinates.

pub fn dv(&self) -> &i8

Returns a reference to the horizontal difference of coordinates.

pub fn dh_mut(&mut self) -> &mut i8

Returns a mutable reference to the vertical difference of coordinates.

pub fn dv_mut(&mut self) -> &mut i8

Returns a mutable reference to the horizontal difference of coordinates.