Module tokyodoves::error

Error variants

A rough sketch of dependency tree:

Error (root of all errors)

BoardError

BoardCreateError: BoardCreateErrorKind

ActionPerformError: ActionPerformErrorKind + Action

MaskShiftError

ActionConvertError

EncodingError: EncodingErrorKind

DecodingError: DecodingErrorKind

GameError

GameRuleCreateError: GameRuleCreateErrorKind

PlayingError: PlayingErrorKind

AnalysisError

ArgsValidationError: ArgsValidationErrorKind

BoardValueMismatch: std::cmp::Ordering

Enums

ActionConvertError

Errors on conversion between Action and string in SSN

ActionPerformErrorKind

Error kinds on performing Action

AnalysisError

Error variants on analysis for games

ArgsValidationErrorKind

Error kinds on validation of arguments

BoardCreateErrorKind

Error kinds on creating Board

BoardError

Errors associated to the Board struct

DecodingErrorKind

Error kinds on conversion of Action from string in SSN

EncodingErrorKind

Error kinds on conversion of Action into string in SSN

Error

Root of all errors in this crate

GameError

Errors associated to Game

GameRuleCreateErrorKind

Error kinds on creating GameRule

PlayingErrorKind

Error kinds on playing games