Struct tokyodoves::Board

pub struct Board { /* private fields */ }

A board of Tokyo Doves based on bitboard techniques.

See the documentation at the top of this crate for quick start examples.

Implementations

impl Board

pub fn new() -> Self

Creates Board at the beginning of the game.

Examples

The following two ways are equivalent:

use tokyodoves::{Board, BoardBuilder};

let board1 = Board::new();
let board2 = BoardBuilder::new().build_unchecked();
assert_eq!(board1, board2);

BoardBuilder provides a variety of ways to create Board.

pub fn perform(&mut self, action: Action) -> Result<(), Error>

Performs the specified action to self.

Errors

Returns Err if performing the specified action is illegal. self will not be changed when this method returns Err.

Examples

use tokyodoves::{Board, Action, Color, Dove, Shift};

let mut board = Board::new();
let action = Action::Put(Color::Red, Dove::A, Shift::new(0, 1));
let result = board.perform(action);
assert!(result.is_ok());

pub fn perform_copied(&self, action: Action) -> Result<Self, Error>

Returns the result of the perform method without changing self.

Errors

Returns Err if performing the specified action is illegal.

Examples

use tokyodoves::{Board, Action, Color, Dove, Shift};

let board = Board::new();
let action = Action::Put(Color::Red, Dove::A, Shift::new(0, 1));
let next_board = board.perform_copied(action)?;

The code above is almost equivalent to the following.

use tokyodoves::{Board, Action, Color, Dove, Shift};

let board = Board::new();
let action = Action::Put(Color::Red, Dove::A, Shift::new(0, 1));
let mut next_board = board; // copied
next_board.perform(action)?;

pub fn perform_bwd(&mut self, action: Action) -> Result<(), Error>

Performs the specified action to self. The action is checked in backward direction.

The action (A) is legal in backward direction if there is such a legal action (B) that performing B to the board after A results in the board before A.

Errors

Returns Err if performing the specified action is illegal in backward direction. self will not be changed when this method returns Err.

Examples

use tokyodoves::{Board, Action, Color, Dove, Shift};

let mut board = Board::new();
let action = Action::Put(Color::Red, Dove::A, Shift::new(-1, 1));
let result = board.perform_bwd(action);
assert!(result.is_ok());

pub fn perform_bwd_copied(&self, action: Action) -> Result<Self, Error>

Returns the result of the perform_bwd method without changing self.

Errors

Returns Err if performing the specified action is illegal in backward direction.

Examples

use tokyodoves::{Board, Action, Color, Dove, Shift};

let board = Board::new();
let action = Action::Put(Color::Red, Dove::A, Shift::new(0, 1));
let next_board = board.perform_bwd_copied(action)?;

The code above is almost equivalent to the following.

use tokyodoves::{Board, Action, Color, Dove, Shift};

let board = Board::new();
let action = Action::Put(Color::Red, Dove::A, Shift::new(0, 1));
let mut next_board = board; // copied
next_board.perform_bwd(action)?;

pub fn perform_unchecked(&mut self, action: Action)

Performs action to self without legality check.

It may return Ok(()) even if the specified action is illegal in both forward and backward directions. In such a case, the resulting board may (or may not) violate the rule (e.g. some dove would be isolated from others). If you are not sure that the action is legal, consider to select one of the following choices:

Call the check_action method or the check_action_bwd method to check legality in advance.

Call the perform method or the perform_bwd method insted, which calls the check_action method or the check_action_bwd method internally.

Note that this method is faster than the methods perform and perform_bwd thanks to absence of legality check.

Panics

Panics if the action is Move or Put, and the target position is more than two squares far from the current 4x4 field (, which are clearly illegal). self will not be changed when this method returns Err.

Examples

use tokyodoves::{Board, Action, Color, Dove, Shift};

let mut board = Board::new();
let action = Action::Put(Color::Red, Dove::A, Shift::new(0, 1));
board.perform_unchecked(action);

pub fn perform_unchecked_copied(&self, action: Action) -> Self

Returns the result of the perform_unchecked method without changing self.

Panics

Panics if the action is Move or Put, and the target position is more than two squares far from the current 4x4 field (, which are clearly illegal).

Examples

use tokyodoves::{Board, Action, Color, Dove, Shift};

let board = Board::new();
let action = Action::Put(Color::Red, Dove::A, Shift::new(0, 1));
let next_board = board.perform_unchecked_copied(action);

The code above is almost equivalent to the following.

use tokyodoves::{Board, Action, Color, Dove, Shift};

let board = Board::new();
let action = Action::Put(Color::Red, Dove::A, Shift::new(0, 1));
let mut next_board = board; // copied
next_board.perform_unchecked(action);

pub fn check_action(&self, action: Action) -> Result<(), Error>

Returns Ok(()) if the given action is legal.

Errors

Returns Err if performing the specified action is illegal.

Examples

use tokyodoves::{Board, Action, Color, Dove, Shift};

let mut board = Board::new();
let action = Action::Put(Color::Red, Dove::A, Shift::new(0, 1));
assert!(board.check_action(action).is_ok());

pub fn check_action_bwd(&self, action: Action) -> Result<(), Error>

Check if backward action is legal in backward direction.

The action (A) is legal in backward direction if there is such a legal action (B) that performing B to the board after A results in the board before A.

Errors

Returns Err if performing the specified action is illegal in backward direction.

Examples

use tokyodoves::{Board, Action, Color, Dove, Shift};

let mut board = Board::new();
let action = Action::Put(Color::Red, Dove::A, Shift::new(-1, 1));
assert!(board.check_action_bwd(action).is_ok());

pub fn legal_actions( &self, player: Color, contains_put: bool, contains_move: bool, contains_remove: bool ) -> ActionsFwd

Collects and returns all legal Actions performed by the specified player.

Three boolean arguments — contains_put, contains_move and contains_remove — specify what kinds of Actions should be contained.

Examples

use tokyodoves::{Board, Color};

let board = Board::new();
let actions = board.legal_actions(Color::Red, true, true, true);
for action in actions {
    println!("{}", board.perform_copied(action)?);
}

pub fn legal_actions_bwd( &self, player: Color, contains_put: bool, contains_move: bool, contains_remove: bool ) -> ActionsBwd

Collects and returns all Actions legal in backward direction performed by the specified player.

Three boolean arguments — contains_put, contains_move and contains_remove — specify what kinds of Actions should be contained.

Examples

use tokyodoves::{Board, Color};

let board = Board::new();
let actions = board.legal_actions_bwd(Color::Red, true, true, true);
for action in actions {
    println!("{}", board.perform_bwd_copied(action)?);
}

pub fn swap_color(&mut self)

Swaps the colors red and green.

Examples

use std::str::FromStr;
use tokyodoves::BoardBuilder;

let mut board1 = BoardBuilder::from_str("BbA")?.build()?;
let board2 = BoardBuilder::from_str("bBa")?.build()?;
board1.swap_color();
assert_eq!(board1, board2);

pub fn position_in_rbcc(&self, player: Color, dove: Dove) -> Option<Shift>

Returns the position of specified player and dove in Red-Boss-Centered Coordinate (RBCC).

dv and dh in RBCC is shown in below (B is red boss):

    dh  -3   -2   -1    0    +1   +2   +3
  dv  +----+----+----+-----+----+----+----+
  -3  |    |    |    |     |    |    |    |
      +----+----+----+-----+----+----+----+
  -2  |    |    |    |     |    |    |    |
      +----+----+----+-----+----+----+----+
  -1  |    |    |    |     |    |    |    |
      +----+----+----+-----+----+----+----+
   0  |    |    |    |  B  |    |    |    |
      +----+----+----+-----+----+----+----+
  +1  |    |    |    |     |    |    |    |
      +----+----+----+-----+----+----+----+
  +2  |    |    |    |     |    |    |    |
      +----+----+----+-----+----+----+----+
  +3  |    |    |    |     |    |    |    |
      +----+----+----+-----+----+----+----+

It returns None if specified dove is not found on the field.

pub fn count_doves_on_field(&self) -> usize

Counts the number of all doves on the field.

Examples

use tokyodoves::Board;

let board = Board::new();
assert_eq!(board.count_doves_on_field(), 2);

pub fn count_doves_on_field_of(&self, player: Color) -> usize

Counts the number of doves of the player on the field.

Examples

use tokyodoves::{Board, Color};

let board = Board::new();
assert_eq!(board.count_doves_on_field_of(Color::Red), 1);

pub fn doves_in_hand_of(&self, player: Color) -> DoveSet

Collects and returns the set of doves in the player’s hand.

Examples

use tokyodoves::{Board, Color};

let board = Board::new();
for dove in board.doves_in_hand_of(Color::Red) {
    println!("{:?}", dove); // A, Y, M, T, H
}

pub fn doves_on_field_of(&self, player: Color) -> DoveSet

Collects and returns the set of doves of the player on the field.

Examples

use tokyodoves::{Board, Color};

let board = Board::new();
for dove in board.doves_on_field_of(Color::Red) {
    println!("{:?}", dove); // B
}

pub fn is_on_field(&self, player: Color, dove: Dove) -> bool

Returns true if the player’s dove is on the field.

Examples

use tokyodoves::{Board, Color, Dove};

let board = Board::new();
assert!(board.is_on_field(Color::Red, Dove::B));
assert!(!board.is_on_field(Color::Green, Dove::A));

pub fn is_in_hand(&self, player: Color, dove: Dove) -> bool

Returns true if the player’s dove is in their hand.

Examples

use tokyodoves::{Board, Color, Dove};

let board = Board::new();
assert!(board.is_in_hand(Color::Green, Dove::A));
assert!(!board.is_in_hand(Color::Red, Dove::B));

pub fn liberty(&self, player: Color, dove: Dove) -> Option<usize>

Returns “liberty” of player’s dove, where liberty is the number of free squares next to the specified dove.

Note that None and Some(0) have different meanings; None means that the specified dove is not on the board while Some(0) means that the specified dove is on the field and completely surrounded.

Examples

+---+---+---+---+
| b | T |   |   |
+---+---+---+---+
|   | B |   |   |
+---+---+---+---+
|   | a | H | M |
+---+---+---+---+
|   |   | m |   |
+---+---+---+---+

In this case, the liberty of “B” is two. The liberty of “T” is one, because three of four squares next to “T” is occupied by “B”, “b” and a wall.

use std::str::FromStr;
use tokyodoves::{BoardBuilder, Color, Dove};

let board = BoardBuilder::from_str("bT; B; aHM;  m")?.build()?;
assert_eq!(Some(2), board.liberty(Color::Red, Dove::B));
assert_eq!(Some(1), board.liberty(Color::Red, Dove::T));
assert_eq!(None, board.liberty(Color::Green, Dove::H)); // not on field

pub fn liberty_of_boss(&self, player: Color) -> usize

Returns “liberty” of player’s boss-hato.

The meaning of liberty is written in the description of the liberty method. A player loses when the liberty of their boss-hato becomes 0.

Panics

Panics if player’s boss is not on the field, which does not happen in normal play.

Examples

+---+---+---+---+
| b | T |   |   |
+---+---+---+---+
|   | B |   |   |
+---+---+---+---+
|   | a | H | M |
+---+---+---+---+
|   |   | m |   |
+---+---+---+---+

use std::str::FromStr;
use tokyodoves::{BoardBuilder, Color, Dove};

let board = BoardBuilder::from_str("bT; B; aHM;  m")?.build()?;
assert_eq!(2, board.liberty_of_boss(Color::Red));
assert_eq!(1, board.liberty_of_boss(Color::Green));

pub fn surrounded_status(&self) -> SurroundedStatus

Returns information about whether bosses are surrounded or not.

See the description of SurroundedStatus for detail.

Examples

use std::str::FromStr;
use tokyodoves::{BoardBuilder, Color, Dove, SurroundedStatus};

let board1 = BoardBuilder::from_str("bBh;Aa m;  M; H Y")?.build()?;
let board2 = BoardBuilder::from_str("b h;AaBm;  M; H Y")?.build()?;
let board3 = BoardBuilder::from_str("b h;Aa m; BM ; H Y")?.build()?;

assert_eq!(SurroundedStatus::Both, board1.surrounded_status());
assert_eq!(
    SurroundedStatus::OneSide(Color::Red),
    board2.surrounded_status()
);
assert_eq!(SurroundedStatus::None, board3.surrounded_status());

pub fn minimum_rectangle(&self) -> Rectangle

Returns the minimum rectangle that contains all doves on the field.

It has fields hmin, hmax, vmin and vmax, standing for coordinates of edges of the rectangle.

Examples

[Case 1]
    0   1   2   3
  +---+---+---+---+
0 | b |   | T |   |
  +---+---+---+---+
1 |   | B | a |   |
  +---+---+---+---+
2 |   |   |   |   |
  +---+---+---+---+
3 |   |   |   |   |
  +---+---+---+---+
-> hmin=0, hmax=2, vmin=0, vmax=1

[Case 2]
    0   1   2   3
  +---+---+---+---+
0 |   |   |   |   |
  +---+---+---+---+
1 |   | b |   |   |
  +---+---+---+---+
2 | a |   | H | B |
  +---+---+---+---+
3 |   |   |   |   |
  +---+---+---+---+
-> hmin=0, hmax=3, vmin=1, vmax=2

[Case 3]
    0   1   2   3
  +---+---+---+---+
0 |   |   |   |   |
  +---+---+---+---+
1 |   |   |   |   |
  +---+---+---+---+
2 |   |   |   |   |
  +---+---+---+---+
3 | h | B | b | Y |
  +---+---+---+---+
-> hmin=0, hmax=3, vmin=3, vmax=3

pub fn to_u64(&self) -> u64

Returns a light expression of a Board in u64.

64 bits consist of the following three parts:

Top 4 bits: empty (filled by 0)

Next 12 bits: each bit means that each kind of dove is on the field (1) or not (0)

Next 48 bits: positions of doves are expressed by 0~15, which consumes 4 bits per one dove (4 bits x 12 doves = 48 bits)

The order of 12 doves is “B > b > A > a > Y > y > M > m > T > t > H > h”, where capital/lower cases mean red/green dove, respectively.

Examples

Initial board:

+---+---+---+---+     +---+---+---+---+
| b |   |   |   |     | 0 | 1 | 2 | 3 |
+---+---+---+---+     +---+---+---+---+
| B |   |   |   |     | 4 | 5 | 6 | 7 |
+---+---+---+---+ <=> +---+---+---+---+
|   |   |   |   |     | 8 | 9 |10 |11 |
+---+---+---+---+     +---+---+---+---+
|   |   |   |   |     |12 |13 |14 |15 |
+---+---+---+---+     +---+---+---+---+

As a code:

use tokyodoves::Board;

let board = Board::new();
let top_4 = 0b0000;
let next_12 = 0b110000_000000; // B + b
let next_48 = 4 << 44 | 0 << 40; // B at 4 + b at 0
let hash = top_4 << 60 | next_12 << 48 | next_48;
assert_eq!(board.to_u64(), hash);

pub fn to_invariant_u64(&self, next_player: Color) -> u64

Returns a light expression of u64 invariant under reflection, rotation and translation transformations.

The composition of the returned value is the same as that of to_u64. This method returns the same value for boards that

coinside with each other under reflection, rotation, translation, and any compositions of them.

coinside with each other after alternating next player and swapping colors of doves simultaneously.

Examples

Boards

+---+---+---+---+
| b | B |   |   |
+---+---+---+---+
|   | H |   |   |
+---+---+---+---+
|   | y |   |   |
+---+---+---+---+
|   |   |   |   |
+---+---+---+---+

and

+---+---+---+---+
|   |   |   |   |
+---+---+---+---+
|   | b |   |   |
+---+---+---+---+
|   | B | H | y |
+---+---+---+---+
|   |   |   |   |
+---+---+---+---+

viewed from red and

+---+---+---+---+
|   |   |   |   |
+---+---+---+---+
|   | B |   |   |
+---+---+---+---+
|   | b | h | Y |
+---+---+---+---+
|   |   |   |   |
+---+---+---+---+

viewed from green return the same values.

use std::str::FromStr;
use tokyodoves::{BoardBuilder, Color};

let board1 = BoardBuilder::from_str("bB; H; y")?.build()?;
let hash1 = board1.to_invariant_u64(Color::Red);
let board2 = BoardBuilder::from_str("; b; BHy")?.build()?;
let hash2 = board2.to_invariant_u64(Color::Red);
let board3 = BoardBuilder::from_str("; B; bhY")?.build()?;
let hash3 = board3.to_invariant_u64(Color::Green);
assert_eq!(hash1, hash2);
assert_eq!(hash2, hash3);

pub fn to_4x4_matrix(&self) -> [[Option<(Color, Dove)>; 4]; 4]

Returns 4x4 matrix (array of array) representing the board.

Examples

use tokyodoves::{Board, Color, Dove};

let board = Board::new();
let matrix = [
    [Some((Color::Green, Dove::B)), None, None, None],
    [Some((Color::Red, Dove::B)), None, None, None],
    [None, None, None, None],
    [None, None, None, None],
];
assert_eq!(board.to_4x4_matrix(), matrix);

pub fn display(&self) -> BoardDisplay<'_>

Returns BoardDisplay to display the board.

BoardDisplay provides a way to configure display styles. See its documentation for more.

pub fn to_framed_string(&self) -> String

Returns a String expression with a frame.

Examples

use tokyodoves::Board;

let board = Board::new();
let string: String = [
    "+---+---+---+---+",
    "| b |   |   |   |",
    "+---+---+---+---+",
    "| B |   |   |   |",
    "+---+---+---+---+",
    "|   |   |   |   |",
    "+---+---+---+---+",
    "|   |   |   |   |",
    "+---+---+---+---+",
].join("\n");
assert_eq!(board.to_framed_string(), string);

pub fn to_simple_string(&self, empty: char, delimiter: &str) -> String

Returns a simple String expression.

empty specifies what character fills empty squares. delimiter specifies what &str divides each row.

Examples

use tokyodoves::Board;

let board = Board::new();
let string = String::from("b---;B---;----;----");
assert_eq!(string, board.to_simple_string('-', ";"));